Move DynamoDB-specific test to aws directory
