+++ pkg/chunk/storage/storage_client_test.go
@@ -5,7 +5,6 @@ import (
 	"math/rand"
 	"sort"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
 
@@ -56,37 +55,5 @@ func TestChunksBasic(t *testing.T) {
 				require.Equal(t, chunksToGet[i].ExternalKey(), chunksWeGot[i].ExternalKey(), strconv.Itoa(i))
 			}
 		}
-	})
-}
-
-type clientWithErrorParameters interface {
-	SetErrorParameters(provisionedErr, errAfter int)
-}
-
-func TestChunksPartialError(t *testing.T) {
-	forAllFixtures(t, func(t *testing.T, client chunk.StorageClient) {
-		// This test is currently very specialised for DynamoDB
-		if !strings.Contains(t.Name(), "DynamoDB") {
-			return
-		}
-		// We use some carefully-chosen numbers:
-		// Start with 150 chunks; DynamoDB writes batches in 25s so 6 batches.
-		// We tell the client to error after 7 operations so all writes succeed
-		// and then the 2nd read fails, so we read back only 100 chunks
-		if ep, ok := client.(clientWithErrorParameters); ok {
-			ep.SetErrorParameters(22, 7)
-		} else {
-			t.Error("DynamoDB test fixture does not support SetErrorParameters() call")
-			return
-		}
-		ctx := context.Background()
-		_, chunks, err := testutils.CreateChunks(0, 150)
-		require.NoError(t, err)
-		err = client.PutChunks(ctx, chunks)
-		require.NoError(t, err)
-
-		chunksWeGot, err := client.GetChunks(ctx, chunks)
-		require.Error(t, err)
-		require.Equal(t, 100, len(chunksWeGot))
 	})
 }
